// src/usecase/customer/create/create.customer.usecase.ts
import CustomerFactory from "../../../domain/customer/factory/customer-factory";
import CustomerRepositoryInterface from "../../../domain/customer/repository/customer-interface";
import Address from "../../../domain/customer/value-object/Address";
import {
  InputCreateCustomerDto,
  OutputCreateCustomerDto,
} from "./create.customer.dto";

export default class CustomerCreateUseCase {
  private customerRepository: CustomerRepositoryInterface;

  constructor(customerRepository: CustomerRepositoryInterface) {
    this.customerRepository = customerRepository;
  }

  async execute(
    input: InputCreateCustomerDto
  ): Promise<OutputCreateCustomerDto> {
    const customer = CustomerFactory.createWithAddress(
      input.name,
      new Address(
        input.address.street,
        input.address.number,
        input.address.zip,
        input.address.city
      )
    );

    await this.customerRepository.create(customer)

    return {
        id: customer.id,
        name: customer.name,
        address: {
            street: customer.Address.street,
            number: customer.Address.number,
            zip: customer.Address.zip,
            city: customer.Address.city
        }
    }
  }
}
